Changes in durch strategy

+++ Mariasek.Engine/PlayerDebugInfo.cs
@@ -31,7 +31,7 @@ namespace Mariasek.Engine
         //public List<int> EstimatedGreaseProbabilityList { get; set; }
         public int EstimatedHundredWinProbability { get; set; }
         public int EstimatedDurchWinProbability { get; set; }
-        public float EstimatedAverageHundredMoneyWon { get; set; }
+        public int EstimatedAverageHundredMoneyWon { get; set; }
         public string ProbDebugInfo { get; set; }
 
         public PlayerDebugInfo()
